requirements.txt
aiogram==3.15.0
python-dotenv==1.0.0
google-generativeai==0.8.3
python-docx==1.1.2
python-pptx==1.0.2
pdfminer.six==20231228
pytz==2024.1
pytesseract==0.3.13
Pillow==11.0.0
requests==2.32.3
beautifulsoup4==4.12.3

fastapi
uvicorn
psutil
